split validators to help avoid circular dependencies
